requirements.txt
nltk==3.6.2
rank_bm25==0.2
streamlit==0.72.0
PyMuPDF==1.18.14
numpy==1.26.4
pandas==1.2.1  
pandas-profiling==3.2.0
ollama==0.4.7 
transformers==4.49.0
sentence-transformers==3.4.1
faiss-cpu==1.10.0
langchain==0.3.19
pymilvus==2.5.4
packaging==24.2
accelerate==1.4
pdfplumber==0.11.5
spacy==3.8.3
onnxruntime==1.19.2